Addition to the UI
Possibility to press escape key to exit the application.
Fixed some unwanted behaviours in the UI, such as interactable calibration button when calibration was already done.
Added the possibility of coming back to the menu in visualisation.
Added an event when the animation finishes. Will be of use later.
Added text in visualisation to get different data frames. Did not implement its script manager yet.